refactor: use actual error handling functions in CLI error tests
Removes duplicated error handling logic in cli-error-handling.test.ts by creating test helpers
that use the same logic as the actual CLI error handler. This ensures that tests verify the
actual implementation rather than a duplicate of it.

- Created test-helpers/cli-handlers.ts with extracted error handling functions
- Updated cli-error-handling.test.ts to use these helpers
- Added proper mocking to prevent unintended side effects
- Updated test assertions to work with the actual implementation

Resolves AC 1.2 (DRY Violation in CLI Error Tests)
